first

+++ plan/views.py
@@ -18,8 +18,8 @@ from django.template import RequestContext
 from django.template.context_processors import csrf
 
 from mysite import settings
-from plan.forms import ReportForm, LoginForm, WorkPartForm, ReportFormPage2
-from plan.models import Worker, WorkReport, WorkPart, StandartWork
+from plan.forms import ReportForm, LoginForm, WorkPartForm, HardwareEquipmentForm, RejectForm
+from plan.models import Worker, WorkReport, WorkPart, StandartWork, HardwareEquipment, Reject
 from plan.workReportGenerator import generateReport
 
 
@@ -88,6 +88,7 @@ def workReportPage1(request, workReport_id):
             work_report.stockMan = form.cleaned_data["stockMan"]
             work_report.adate = form.cleaned_data["adate"]
             work_report.VIKDate = form.cleaned_data["VIKDate"]
+            work_report.note = form.cleaned_data["note"]
             work_report.save()
 
             # возвращаем простое окно регистрации
@@ -100,6 +101,7 @@ def workReportPage1(request, workReport_id):
                     'worker': form.cleaned_data["worker"],
                     'stockMan': form.cleaned_data["stockMan"],
                     'date': form.cleaned_data["date"],
+                    'note': form.cleaned_data["note"],
                     }
             return render(request, "plan/workReportPage1.html", {
                 'form': ReportForm(data),
@@ -119,17 +121,23 @@ def workReportPage2(request, workReport_id):
             super(RequiredFormSet, self).__init__(*args, **kwargs)
             for form in self.forms:
                 form.empty_permitted = False
+
     wr = WorkReport.objects.get(pk=workReport_id)
 
-  #  if wr.workPart.all().exists():
+    #  if wr.workPart.all().exists():
 
     ReportFormset = formset_factory(WorkPartForm, max_num=10, formset=RequiredFormSet)
     if request.method == 'POST':
         report_formset = ReportFormset(request.POST, request.FILES)
         if report_formset.is_valid():
-            wr.workPart.all().delete()
+            if wr.workPart.all().exists():
+                wr.workPart.all().delete()
             for form in report_formset.forms:
-                w = WorkPart.create(form.cleaned_data)
+                w = WorkPart.objects.create(startTime=form.cleaned_data["startTime"],
+                                            endTime=form.cleaned_data["endTime"],
+                                            standartWork=form.cleaned_data["standartWork"],
+                                            workPlace=None,
+                                            rationale=None)
                 w.save()
                 wr.workPart.add(w)
                 print (form.cleaned_data)
@@ -155,24 +163,192 @@ def workReportPage2(request, workReport_id):
 
 
 def workReportPage3(request, workReport_id):
-    wRep = WorkReport.objects.get(pk=workReport_id)
+    class RequiredFormSet(BaseFormSet):
+        def __init__(self, *args, **kwargs):
+            super(RequiredFormSet, self).__init__(*args, **kwargs)
+            for form in self.forms:
+                form.empty_permitted = False
 
-    ArticleFormSet = formset_factory(WorkPartForm)
+    wr = WorkReport.objects.get(pk=workReport_id)
+
+    #  if wr.workPart.all().exists():
+
+    ReportFormset = formset_factory(WorkPartForm, max_num=10, formset=RequiredFormSet)
     if request.method == 'POST':
-        formset = ArticleFormSet(request.POST, request.FILES)
-        if formset.is_valid():
-            # do something with the formset.cleaned_data
-            pass
+        report_formset = ReportFormset(request.POST, request.FILES)
+        if report_formset.is_valid():
+            if wr.factWorkPart.all().exists():
+                wr.factWorkPart.all().delete()
+            for form in report_formset.forms:
+                w = WorkPart.objects.create(startTime=form.cleaned_data["startTime"],
+                                            endTime=form.cleaned_data["endTime"],
+                                            standartWork=form.cleaned_data["standartWork"],
+                                            workPlace=None,
+                                            rationale=None)
+                w.save()
+                wr.factWorkPart.add(w)
+                print (form.cleaned_data)
+            return HttpResponseRedirect('/workReport/page4/' + str(workReport_id) + '/')
     else:
         data = {
-            'form-TOTAL_FORMS': u'6',
-            'form-INITIAL_FORMS': u'1',
-            'form-MAX_NUM_FORMS': u'10',
+            'form-TOTAL_FORMS': '3',
+            'form-INITIAL_FORMS': '0',
+            'form-MAX_NUM_FORMS': '',
+            'form-0-startTime': '08:30',
+            'form-0-endTime': '08:45',
+            'form-1-startTime': '16:30',
+            'form-1-endTime': '16:45',
+            'form-2-startTime': '16:45',
+            'form-2-endTime': '17:00',
+            'form-0-standartWork': StandartWork.objects.get(text='Получение наряда и ТМЦ для выполнения работ'),
+            'form-1-standartWork': StandartWork.objects.get(
+                text='Отчет о выполненных работах перед РП и ознакомление со сменным нарядом на следующий рабочий день, проверка инструмента, сдача ТМЦ'),
+            'form-2-standartWork': StandartWork.objects.get(text='Уборка рабочего места')
         }
-        formset = ArticleFormSet(data)
+        report_formset = ReportFormset(data)
+    c = {'report_formset': report_formset,
+         'caption': 'Фактически выполненные работы'
+         }
+    #  c.update(csrf(request))
+    return render(request, 'plan/workReportFormset.html', c)
 
-    return render_to_response('plan/workReportFormset.html', {'formset': formset})
 
+def workReportPage4(request, workReport_id):
+    class RequiredFormSet(BaseFormSet):
+        def __init__(self, *args, **kwargs):
+            super(RequiredFormSet, self).__init__(*args, **kwargs)
+            for form in self.forms:
+                form.empty_permitted = False
+
+    wr = WorkReport.objects.get(pk=workReport_id)
+
+
+    #  if wr.workPart.all().exists():
+
+    ReportFormset = formset_factory(HardwareEquipmentForm, max_num=10, formset=RequiredFormSet)
+    if request.method == 'POST':
+        report_formset = ReportFormset(request.POST, request.FILES)
+        if report_formset.is_valid():
+            if wr.planHardware.all().exists():
+                wr.planHardware.all().delete()
+            for form in report_formset.forms:
+                w = HardwareEquipment.objects.create(
+                    # деталь
+                    equipment=form.cleaned_data["equipment"],
+                    material=form.cleaned_data["material"],
+                    usedCnt=form.cleaned_data["usedCnt"],
+                    getCnt=form.cleaned_data["getCnt"],
+                    rejectCnt=form.cleaned_data["rejectCnt"],
+                    dustCnt=form.cleaned_data["dustCnt"],
+                    remainCnt=form.cleaned_data["remainCnt"]
+                )
+                w.save()
+                wr.planHardware.add(w)
+                print (form.cleaned_data)
+            return HttpResponseRedirect('/workReport/page5/' + str(workReport_id) + '/')
+    else:
+        data = {
+            'form-TOTAL_FORMS': '0',
+            'form-INITIAL_FORMS': '0',
+            'form-MAX_NUM_FORMS': '',
+        }
+        report_formset = ReportFormset(data)
+    c = {'report_formset': report_formset,
+         'caption': 'Плановая выдача оборудования'
+         }
+    #  c.update(csrf(request))
+    return render(request, 'plan/workReportFormset.html', c)
+
+
+def workReportPage5(request, workReport_id):
+    class RequiredFormSet(BaseFormSet):
+        def __init__(self, *args, **kwargs):
+            super(RequiredFormSet, self).__init__(*args, **kwargs)
+            for form in self.forms:
+                form.empty_permitted = False
+
+    wr = WorkReport.objects.get(pk=workReport_id)
+
+    #  if wr.workPart.all().exists():
+
+    ReportFormset = formset_factory(HardwareEquipmentForm, max_num=10, formset=RequiredFormSet)
+    if request.method == 'POST':
+        report_formset = ReportFormset(request.POST, request.FILES)
+        if report_formset.is_valid():
+            if wr.noPlanHardware.all().exists():
+                wr.noPlanHardware.all().delete()
+            for form in report_formset.forms:
+                w = HardwareEquipment.objects.create(
+                    # деталь
+                    equipment=form.cleaned_data["equipment"],
+                    material=form.cleaned_data["material"],
+                    usedCnt=form.cleaned_data["usedCnt"],
+                    getCnt=form.cleaned_data["getCnt"],
+                    rejectCnt=form.cleaned_data["rejectCnt"],
+                    dustCnt=form.cleaned_data["dustCnt"],
+                    remainCnt=form.cleaned_data["remainCnt"]
+                )
+                w.save()
+                wr.noPlanHardware.add(w)
+                print (form.cleaned_data)
+            return HttpResponseRedirect('/workReport/page6/' + str(workReport_id) + '/')
+    else:
+        data = {
+            'form-TOTAL_FORMS': '0',
+            'form-INITIAL_FORMS': '0',
+            'form-MAX_NUM_FORMS': '',
+        }
+        report_formset = ReportFormset(data)
+    c = {'report_formset': report_formset,
+         'caption': 'Внеплановая выдача оборудования'
+         }
+    #  c.update(csrf(request))
+    return render(request, 'plan/workReportFormset.html', c)
+
+
+def workReportPage6(request, workReport_id):
+    class RequiredFormSet(BaseFormSet):
+        def __init__(self, *args, **kwargs):
+            super(RequiredFormSet, self).__init__(*args, **kwargs)
+            for form in self.forms:
+                form.empty_permitted = False
+
+    wr = WorkReport.objects.get(pk=workReport_id)
+
+    #  if wr.workPart.all().exists():
+
+    ReportFormset = formset_factory(RejectForm, max_num=10, formset=RequiredFormSet)
+    if request.method == 'POST':
+        report_formset = ReportFormset(request.POST, request.FILES)
+        if report_formset.is_valid():
+            if wr.rejected.all().exists():
+                wr.rejected.all().delete()
+            for form in report_formset.forms:
+                w = Reject.objects.create(
+                    # деталь
+                    equipment=form.cleaned_data["equipment"],
+                    material=form.cleaned_data["material"],
+                    cnt=form.cleaned_data["cnt"],
+                )
+                w.save()
+                wr.rejected.add(w)
+                print (form.cleaned_data)
+            return HttpResponseRedirect('/workReport/page7/' + str(workReport_id) + '/')
+    else:
+        data = {
+            'form-TOTAL_FORMS': '0',
+            'form-INITIAL_FORMS': '0',
+            'form-MAX_NUM_FORMS': '',
+        }
+        report_formset = ReportFormset(data)
+    c = {'report_formset': report_formset,
+         'caption': 'Направлено в изолятор брака'
+         }
+    #  c.update(csrf(request))
+    return render(request, 'plan/workReportFormset.html', c)
+
+def workReportPage7(request, workReport_id):
+    return render(request, 'plan/workReportFinal.html', {'id':workReport_id})
 
 def workersView(request):
     return None
@@ -200,5 +376,51 @@ def createWorkReport(request):
     return HttpResponseRedirect('/workReport/page1/' + str(w.pk) + '/')
 
 
-def printReport(request):
-    return render_to_response('plan/test.html')
+def printReport(request,workReport_id):
+    rationales = [
+        ['2,3', 'Я так захотел'],
+        ['3,4,5', 'А это заставили']
+    ]
+    works = [
+        ['1', '-', 'Получение наряда и ТМЦ для выполнения работ', '08:30', '08:45'],
+        ['2', '1.2.3', 'Изготовление деталей оснастки для сушки лейнеров по прилагаемому чертежу в кол-ве одного '
+                       'комплекта', '13:15', '16:15'],
+        ['3', '-', 'Отчет о выполненных работах перед РП и ознакомление со сменным нарядом на следующий рабочий день,'
+                   'проверка инструмента, сдача ТМЦ', '16:15', '16:45'],
+        ['4', '1.2.3', 'Уборка рабочего места', '16:45', '17:00'],
+    ]
+
+    factWorks = [
+        ['1', '-', 'Получение наряда и ТМЦ для выполнения работ', '08:30', '08:45'],
+        ['2', '1.2.3', 'Изготовление деталей оснастки для сушки лейнеров по прилагаемому чертежу в кол-ве одного '
+                       'комплекта', '13:15', '16:15'],
+        ['3', '-', 'Отчет о выполненных работах перед РП и ознакомление со сменным нарядом на следующий рабочий'
+                   'день, проверка инструмента, сдача ТМЦ ', '16:15', '16:45'],
+        ['4', '1.2.3', 'Уборка рабочего места', '16:45', '17:00'],
+    ]
+    note = '''Примечание 1 (обязательное):
+       Максимальный срок проведения ВИК (входного контроля) до конца рабочего дня 16.01.2017 г.'''
+
+    planEquipment = [
+        ['Перчатки х/б', '123', 'пара', '1', '0', '0', '0', '-1'],
+        ['Пруток бронза', 'Б132r', 'мм', '200', '0', '0', '0', '-1'],
+    ]
+    nonPlanEquipment = [
+        ['asf х/б', '123', 'пара', '1', '0', '0', '0', '-1'],
+        ['Пруток бронза', 'Б132r', 'мм', '200', '1', '2', '1', '5'],
+    ]
+    dust = [
+        ['хлам 1', '100'],
+        ['хлам 2', '500']
+    ]
+
+    document = generateReport('ШАВ', 'Шанин А.В.', 'Бука А.В', '124', "Головнёв А.К.", datetime.date.today(), 'Токарь',
+                              rationales, works, factWorks, 'Шанин А.В.', 'Шанин А.В.', 'Хионин Б.Г.', note,
+                              'аттестация отутствует', dust, planEquipment, nonPlanEquipment)
+
+    response = HttpResponse(content_type='application/vnd.openxmlformats-officedocument.wordprocessingml.document')
+    response['Content-Disposition'] = 'attachment; filename=report.docx'
+    document.save(response)
+
+    return response
+
